LayoutTests/media/W3C/video/events/event_pause_manual-expected.txt

calling play() then pause() on non-autoplay video should trigger pause event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS true is true

TEST COMPLETE
spec reference
